refactor(db): extract analytics signals
